requirements.txt
dnspython==1.12.0
dyn==1.4.2
tld==0.10
